Fixes after first test

+++ setup.sh
@@ -1,15 +1,25 @@
 #!/bin/bash
 
+GREEN='\033[0;32m'
+LGREEN='\033[1;32m'
+YELLOW='\033[1;33m'
+NC='\033[0m' # No Color
+
 step () {
-    echo
-    echo
-    echo "Next step: $1"
+    echo -e "${GREEN}-------------${NC}"
+    echo -e "${LGREEN}Next step: ${YELLOW}$1${NC}"
+    echo -e "${GREEN}-------------${NC}"
     echo "<Press enter to continue>"
     read
+    echo
 }
 
+if [ "$EUID" -e 0 ]
+  then echo "Please don't run as root"
+  exit
+fi
+
 sudo mkdir -p ~/temp
-cd ~/temp
 
 # Packages for fetching
 sudo apt install wget curl
@@ -35,19 +45,17 @@ echo "[Seat:*]" | sudo tee -a /etc/lightdm/lightdm.conf.d/01_my.conf
 echo "greeter-hide-users=false" | sudo tee -a /etc/lightdm/lightdm.conf.d/01_my.conf
 
 step "Simple LightDM GTK theme"
-[ -d "./simple-lightdm-gtk-theme" ] && sudo rm -rf ./simple-lightdm-gtk-theme
-git clone https://github.com/freande/simple-lightdm-gtk-theme.git
-cd simple-lightdm-gtk-theme
-sudo chmod +x install.sh
-sudo ./install.sh
-cd ..
+[ -d "~/temp/simple-lightdm-gtk-theme" ] && sudo rm -rf ~/temp/simple-lightdm-gtk-theme
+git clone https://github.com/freande/simple-lightdm-gtk-theme.git ~/temp/simple-lightdm-gtk-theme
+sudo chmod +x ~/temp/simple-lightdm-gtk-theme/install.sh
+sudo ~/temp/simple-lightdm-gtk-theme/install.sh
 
 step "Remove Grub timeout"
 sudo sed -i 's/^GRUB_TIMEOUT=[[:digit:]]*$/GRUB_TIMEOUT=0/g' /etc/default/grub
 sudo update-grub
 
 step "Desktop packages"
-sudo apt install awesome rofi zsh zplug picom thunar neovim policykit-1 lxpolkit yad pulseaudio pavucontrol upower connman bluez
+sudo apt install awesome rofi zsh zplug picom thunar neovim lxpolkit yad pulseaudio pavucontrol upower connman
 ## maybe pulseaudio-module-bluetooth instead of pulseaudio?
 
 step "Copy config"
@@ -56,23 +64,18 @@ sudo ./copy-config.sh
 
 step "AwesomeWM deps"
 sudo apt install luarocks
-luarocks install pulseaudio_dbus
+sudo luarocks install pulseaudio_dbus
 ## make sure load-module module-dbus-protocol is present in ~/.config/pulse/default.pa
-luarocks install pulseaudio_widget
-luarocks install connman_dbus
-luarocks install connman_widget
-luarocks install upower_dbus
-luarocks install power_widget
+sudo luarocks install pulseaudio_widget
+sudo luarocks install connman_dbus
+sudo luarocks install connman_widget
+sudo luarocks install upower_dbus
+sudo luarocks install power_widget
 
 step "Hack NF font"
-mkdir temp
-cd temp
-git clone https://github.com/ryanoasis/nerd-fonts.git
-cd nerd-fonts
-## Chmod?
-./install.sh Hack
-sudo fc-cache -fv ## Needed?
-cd ..
+wget -P ~/temp/ https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/Hack.zip
+unzip ~/temp/Hack.zip -d /usr/share/fonts/
+sudo fc-cache -fv
 
 step "Alacritty"
 nix-env -iA nixpkgs.alacritty
@@ -82,10 +85,8 @@ chsh -s $(which zsh)
 nix-env -iA nixpkgs.starship
 
 step "Starship config"
-git clone https://github.com/freande/starship-powerline-config.git
-cd starship-powerline-config
-cp ./starship.toml ~/.config/starship.toml
-cd ..
+git clone https://github.com/freande/starship-powerline-config.git ~/temp/starship-powerline-config
+cp ~/temp/starship-powerline-config/starship.toml ~/.config/starship.toml
 
 step "Neovim config (NvChad)"
 nix-env -iA nixpkgs.ripgrep
@@ -97,8 +98,8 @@ export NIXPKGS_ALLOW_UNFREE=1
 nix-env -iA nixpkgs.vscode
 
 step "Chrome (Should really switch to a better browser...)"
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo apt install ./google-chrome-stable_current_amd64.deb
+wget -P ~/temp/ https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+sudo apt install ~/temp/google-chrome-stable_current_amd64.deb
 
 step "UFW"
 sudo apt install ufw
